Better check installed apps

+++ install.sh
@@ -89,7 +89,10 @@ function setup_yunohost() {
 
 function setup_yunorunner() {
     echo_bold "> Installation of YunoRunner..."
-    yunohost app install --force https://github.com/YunoHost-Apps/yunorunner_ynh -a "domain=$domain&path=/$ci_path"
+    if ! yunohost app list --output-as json --quiet | jq -e '.apps[] | select(.id == "yunorunner")' >/dev/null
+    then
+        yunohost app install --force https://github.com/YunoHost-Apps/yunorunner_ynh -a "domain=$domain&path=/$ci_path"
+    fi
     port=$(yunohost app setting yunorunner port)
 
     # Stop YunoRunner
@@ -136,7 +139,8 @@ EOF
 }
 
 function setup_lxd() {
-    if ! yunohost app list | grep -q 'id: lxd'; then
+    if ! yunohost app list --output-as json --quiet | jq -e '.apps[] | select(.id == "lxd")' >/dev/null
+    then
         yunohost app install --force https://github.com/YunoHost-Apps/lxd_ynh
     fi
